+++ src/primitive/defs.rs
@@ -3364,6 +3364,7 @@ impl_primitive!(
     (2[1], SplitByScalar),
     (2[1], SplitBy),
     (2[1], SplitByKeepEmpty),
+    (2, AbsComplex),
     (2, MatrixDiv),
     // Implementation details
     (1, Utf16),